Allow floats in geo lon, lat and radius.
Not sure why this wasn't floats from the start.

+++ tiddlywebplugins/mysql2.py
@@ -358,7 +358,7 @@ class Producer(object):
                 # proximity search on geo.long, geo.lat based on
                 # http://cdent.tiddlyspace.com/bags/cdent_public/tiddlers/Proximity%20Search.html
                 try:
-                    lat, long, radius = [int(item)
+                    lat, long, radius = [float(item)
                             for item in value.split(',', 2)]
                 except ValueError, exc:
                     raise StoreError(
